feat(Exam): CUD based on Matrix and Matrix Detail, still not have GET and GET all with pagination

+++ BackEnd/LIP.Application/CQRS/Handler/Exam/ExamDeleteCommandHandler.cs
@@ -1,10 +1,12 @@
 using LIP.Application.CQRS.Command.Exam;
+using LIP.Application.CQRS.Query.Exam;
+using LIP.Application.DTOs.Response.Exam;
 using LIP.Application.Interface.Repository;
 using MediatR;
 
 namespace LIP.Application.CQRS.Handler.Exam
 {
-    public class ExamDeleteCommandHandler : IRequestHandler<ExamDeleteCommand, bool>
+    public class ExamDeleteCommandHandler : IRequestHandler<ExamDeleteCommand, ExamDeleteResponse>
     {
         private readonly IExamRepository _examRepository;
 
@@ -13,9 +15,39 @@ namespace LIP.Application.CQRS.Handler.Exam
             _examRepository = examRepository;
         }
 
-        public async Task<bool> Handle(ExamDeleteCommand request, CancellationToken cancellationToken)
+        public async Task<ExamDeleteResponse> Handle(ExamDeleteCommand request, CancellationToken cancellationToken)
         {
-            return await _examRepository.DeleteAsync(request);
+            var exam = await _examRepository.GetAsync(new ExamGetQuery { ExamId = request.ExamId });
+            if (exam == null)
+            {
+                return new ExamDeleteResponse
+                {
+                    IsSuccess = false,
+                    Message = "Exam is not found!"
+                };
+            }
+            var rs = await _examRepository.DeleteAsync(request);
+            if (rs)
+            {
+                return new ExamDeleteResponse
+                {
+                    IsSuccess = true,
+                    Message = "Delete Exam successfully!"
+                };
+            }
+            else
+            {
+                return new ExamDeleteResponse
+                {
+                    IsSuccess = false,
+                    Message = "Delete Exam failed!"
+                };
+            }
         }
+
+        //public async Task<bool> Handle(ExamDeleteCommand request, CancellationToken cancellationToken)
+        //{
+        //    return await _examRepository.DeleteAsync(request);
+        //}
     }
 }
